Promote Player management at theatre level
Player managment were done at the scene level which force a lot of
unecessary checks and attributes transfere between the scene switches.
The player is now global to the game. The scene only contains a
PlayerSlot which is an intermediate object to store the z position in
the scene. It contains the reference to the main player.

- Sound shooter temporarily disabled to be promoted as well